pyproject.toml
[project]
name = "senticast-copilot"
version = "0.1.0"
description = "Mineral price forecasting using multi-horizon diffusion models and news sentiment"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    # Deep learning
    "torch>=2.0.0",
    "einops>=0.6.1",
    # Data / numerics
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    # Visualisation
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # NLP / embeddings
    "sentence-transformers>=2.2.0",
    "transformers>=4.30.0",
    # Utilities
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    # Web UI
    "gradio>=4.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ─────────────────────────────────────────────
# uv settings
# ─────────────────────────────────────────────
[tool.uv]
# Recommended: install PyTorch for CPU-only to keep the environment small.
# For GPU (CUDA 12.x) remove these overrides and let torch auto-detect.
# To use CPU-only torch: uv sync --extra-index-url https://download.pytorch.org/whl/cpu
index-url = "https://pypi.org/simple"

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url  = "https://download.pytorch.org/whl/cpu"
explicit = true

# ─────────────────────────────────────────────
# Scripts
# ─────────────────────────────────────────────
[project.scripts]
senticast-train = "src.train:main"
senticast-eval  = "src.evaluate:main"
senticast-app   = "app:main"
